Move unbound exporter code into a package
The main function was feeling a bit big, and this code already has a reasonable
public API.

The Dockerfile had an incorrect build command which needed to be fixed
once it was building more than one thing.
