+++ Dockerfile
@@ -11,7 +11,7 @@ COPY . .
 
 ENV CGO_ENABLED=0
 
-RUN GOOS=$TARGETOS GOARCH=$TARGETPLATFORM go build -v -o /go/bin/unbound_exporter ./...
+RUN GOOS=$TARGETOS GOARCH=$TARGETPLATFORM go build -v -o /go/bin/unbound_exporter .
 
 FROM gcr.io/distroless/static-debian12
